backend/requirements.txt
Flask>=3.0,<4.0
flask-cors>=4.0,<5.0
pdfplumber>=0.11,<0.12
python-docx>=1.1,<2.0
Pillow>=10.0,<12.0
pypdfium2>=4.30,<5.0
pytesseract>=0.3.10,<0.4
